+++ src/wrappers/detritus.ts
@@ -1,9 +1,9 @@
-import { Manager as BaseManager } from "./lib/Manager";
-import type { ManagerOptions, LavalinkNodeOptions } from "./lib/Types";
+import { Manager as BaseManager } from "../lib/Manager";
+import type { ManagerOptions, LavalinkNodeOptions } from "../lib/Types";
 
 import type { ClusterClient, ShardClient } from "detritus-client";
 
-export * from "./index";
+export * from "../index";
 
 export class Manager extends BaseManager {
     public constructor(public readonly client: ClusterClient | ShardClient, nodes: LavalinkNodeOptions[], options: ManagerOptions) {
